Prevent GUI from freezing & provide better structure
Initialization of the start() method in WebcamApplication takes a lots
of time, which freezes the GUI. Put it in another thread and show
message in the GUI to provide better user experience.
A relative change is that the WebcamApplication is created inside
GuiController, not passed in from the outside anymore. So the folder
structure also changes.
Execute main.py instead of alpha.py from now on.

Have all PageController communicate with others, especially main
contoller, through siganls. This hides the concrete widget details,
which mean Pages can be modified independently as long as the signals
remain unchanged.