Fix model not trained bug
What's the bug?
The button-clicked-signal also passes a boolean argument. This make the
parameter (epochs) of the method WritingModelTrainer.train_model get a
unexpected False value. The False value is regarded as 0, so the training
doesn't really do anything.

What I've done:
The bug is simply fixed by adding a decorator to the train_model method.
But since training is a long-running task, it freezes the GUI. So I move
the training method into a worker thread to avoid this problem.

Now, training can be done successfully and confortly smoothly.