Fix bug of tailor-made EAR maker
In the previous commits, the sample ratio list is sorted in-place.
Which means when a new sample is appended, the old sample that being
poped from left is always the smallest EAR value.
The values held by sample ratio list isn't the recent samples but the
largest samples.
Leads to the result that tailor-made EAR kept growing.