Rearrange folder structure
There are roughly 5 parts of applications:
  - blink
  - brightness
  - distance
  - posture
  - time
To make each of them clear, I split lib/ into 5 folders, each of them
corresponds to the applications above.
Detector/calculator/... and guard of the same application are put
together under the folder.

There are 2 scenarios in the PostureGuard:
  - one is calculated in angle
  - another is predicted with model
The algorithm of the latter was directly exposed in the PostureGuard.
I move it out and make it a new class called PosturePredictor.