Have look back wait for real time
To solve the likely good interval problem mentioned in issue #3.
When a LOOK_BACK interval is gotten from the queue, it is first put
aside. The dispatcher method gets another REAL_TIME interval from the
queue, which is likely good interval. The LOOK_BACK one is graded
only if the REAL_TIME one is truly good.

The rule of updating last end time is so adjusted.