Collect and plot centroids of face
Currently we have 3 indices to tell whether the user is concentrating
or not, which are the followings:
  - posture, especially face angle
  - blink rate
  - existence of face
But I find them not "honest" enough to tell the truth. So in this
commit, I try to introduce the concept of the position of face:
when concentrating, people usually won't have their face moving around.