Run the applications in different threads
We now have 4 apps and I though we'll have to wait for all of them to
finish before we go into the next loop, which needs a barrier to do so.
But surprisingly, everything works well without waiting.
Originally, the 4 apps take approximately 0.12s to finish a loop.
Now, only 0.07s is required, which is lot more faster.