Fix posture detection freeze on mtcnn
This problem is that we use TaskWorkers to run applications but
enter the next main while loop without waiting for any of them.
So if the task takes too much time, it will be aborted in the next loop
and immediately starts a new one. MTCNN takes more time than a
non-waiting loop, so it's aborted every single loop and seems like
doing nothing.
To fix this, I use a Barrier to have the main loop wait for the tasks.