Handle exception of no face center in an interval
Originally, when there isn't any faces (the user isn't in front of the
screen) in an interval, we still fit on empty points and calculate the
center of them, which raises a ZeroDivisionError.
Now, this exception is caught and sufficiently bad distance and ratio
are made to lead to an overall low concentration grade without
crashing the application program.

Resolves: #30