Fix grade not plotted problem
We were to have an extra "screen" label out of ColumnHeader's
realization, which causes many problems of inconsistency.
I loose the contraint of ColumnHeader to allow missing label, so
`grade` and `screenshot` can be updated to Monitor with both of
them under the realization of ColumnHeader.
Also there's no more flexibility on SQL commands, since we don't have
"screen" store into the database and its inconsistency with header
causes problem.