Fix col mismatch when insert screenshot-only row
If a screenshot introduce a new student id, then the update of
screenshot creates a new row, but we didn't tell how we're going
to do with all other blank columns, such as id, time and grade.
That made the screenshot show up in the 1st (index 0) column and break
the structure of monitor.
Also refactoring at handling the 1st screenshot comparison.